pyproject.toml
[project]
name = "test-led"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "opencv-python>=4.11.0.86",
    "pillow>=11.2.1",
    "pygame>=2.6.1",
]
